antigravity/code_tracker/active/no_repo/2561a25c79f2a8b50be919266d222ec6_requirements.txt
Œstreamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0
schedule>=1.2.0
{*cascade08{Œ*cascade082,file:///c:/Users/Tomi/FOREX/requirements.txt